+++ src/entities/comment/ui/Comment.tsx
@@ -1,6 +1,6 @@
 import { Dialog, DialogContent, DialogHeader, DialogTitle, Textarea, Button } from "../../../shared/ui"
 import { Edit2, Plus, ThumbsUp, Trash2 } from "lucide-react"
-import { CommentAddDTO, Comment } from "../model"
+import { CommentAddDTO, Comment } from "../model/types"
 import { useComments } from "../api/queries"
 import { useAddComment, useUpdateComment, useLikeComment, useDeleteComment } from "../api"
 import { Post } from "../../post/model"
@@ -41,10 +41,14 @@ export const Comments = ({
           <div key={comment.id} className="flex items-center justify-between text-sm border-b pb-1">
             <div className="flex items-center space-x-2 overflow-hidden">
               <span className="font-medium truncate">{comment.user.username}:</span>
-              <span className="truncate">{highlightText(comment.body, searchQuery)}</span>
+              <span className="block w-70 truncate">{highlightText(comment.body, searchQuery)}</span>
             </div>
             <div className="flex items-center space-x-1">
-              <Button variant="ghost" size="sm" onClick={() => likeComment.mutate({commentId: comment.id, likes: comment.likes})}>
+              <Button
+                variant="ghost"
+                size="sm"
+                onClick={() => likeComment.mutate({ commentId: comment.id, likes: comment.likes })}
+              >
                 <ThumbsUp className="w-3 h-3" />
                 <span className="ml-1 text-xs">{comment.likes}</span>
               </Button>
@@ -140,7 +144,7 @@ export const UpdateCommentDialog = ({
   const updateComment = useUpdateComment(postId)
 
   const handleUpdateComment = () => {
-    if(selectedComment === null) return
+    if (selectedComment === null) return
     updateComment.mutate(selectedComment)
     setShowEditCommentDialog(false)
   }
